Give Request logger its own config
Move the Request logging config into separate config class
